CPU: Fixed ADC sum calculation issue
Previously the sum was being casted to a byte. While this should be done
to set the accumulator, the sum in integer form is still required to
accurately set the carry flag.

Now the sum result is stored as an integer and is only casted during the
setting of the accumulator.

+++ Source/NEST/Classes/CPU.cs
@@ -1025,7 +1025,7 @@ namespace NEST.Classes
             int originalValue = accumulator;
             int additionByte = indexedIndirect(readImmediateByte());
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
 
@@ -1042,7 +1042,7 @@ namespace NEST.Classes
             int originalValue = accumulator;
             int additionByte = readCPURam(readImmediateByte());
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
 
@@ -1100,7 +1100,7 @@ namespace NEST.Classes
             int originalValue = accumulator;
             int additionByte = readImmediateByte();
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);   
+            int sum = originalValue + additionByte + carryAmount;   
             
             accumulator = (byte)(sum & 0xFF);
 
@@ -1153,7 +1153,7 @@ namespace NEST.Classes
             int originalValue = accumulator;
             int additionByte = readCPURam(readImmediateUShort());
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
 
@@ -1218,7 +1218,7 @@ namespace NEST.Classes
             int originalValue = accumulator;
             int additionByte = indirectIndexed(readImmediateByte());
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
 
@@ -1235,7 +1235,7 @@ namespace NEST.Classes
             int originalValue = accumulator;
             int additionByte = zeroPageIndexed(readImmediateByte(), xAddress);
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
 
@@ -1292,7 +1292,7 @@ namespace NEST.Classes
             address += yAddress;
             int additionByte = readCPURam(address);
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
 
@@ -1311,7 +1311,7 @@ namespace NEST.Classes
             address += xAddress;
             int additionByte = readCPURam(address);
             int carryAmount = getFlagStatus(Carry_Flag) ? 1 : 0;
-            int sum = (byte)(originalValue + additionByte + carryAmount);
+            int sum = originalValue + additionByte + carryAmount;
 
             accumulator = (byte)(sum & 0xFF);
